# rust-core/openapi.yaml
openapi: 3.0.3
info:
  title: MantisDB Admin API
  description: |
    Comprehensive REST API for MantisDB administration, monitoring, and management.
    
    ## Features
    - **Health & Monitoring**: Real-time metrics, stats, and system health
    - **Authentication**: Secure JWT-based authentication with RBAC
    - **Tables**: Full CRUD operations on database tables
    - **Queries**: SQL query execution with history tracking
    - **RLS**: Row-Level Security policy management
    - **Logs**: Searchable logs with streaming support
    - **Backups**: Automated backup and restore operations
    - **Config**: Runtime configuration management
    - **WebSocket/SSE**: Real-time event streams
    
  version: 1.0.0
  contact:
    name: MantisDB Team
    url: https://github.com/mantisdb
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8081
    description: Local development server
  - url: http://localhost:8080
    description: API server
...
paths:
...
  /api/tables/{table}:
    get:
      tags: [Tables]
      summary: Get table data
      description: Retrieve all data from a specific table
      security:
        - bearerAuth: []
      parameters:
        - name: table
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Table data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableData'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TableData:
      type: object
      properties:
        rows:
          type: array
          items:
            type: object
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
...